requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
pytesseract==0.3.10
pytz==2024.1
requests==2.31.0
schedule==1.2.1
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.3
